refactor: move NewsletterForm and NewsletterFormWrap
