+++ src/components/GenericPagesComponents/NewsletterForm/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./NewsletterForm";
 export { default as useNewsletterForm } from "./useNewsletterForm";
+export type { NewsletterFormProps } from "./NewsletterForm";
