+++ src/pages/index.tsx
@@ -15,8 +15,8 @@ import useAnalytics from "@/context/Analytics/useAnalytics";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
 import getPageProps from "@/node-lib/getPageProps";
 import PostList from "@/components/SharedComponents/PostList";
-import { useNewsletterForm } from "@/components/Forms/NewsletterForm";
-import NewsletterFormWrap from "@/components/Forms/NewsletterForm/NewsletterFormWrap";
+import { useNewsletterForm } from "@/components/GenericPagesComponents/NewsletterForm";
+import NewsletterFormWrap from "@/components/GenericPagesComponents/NewsletterFormWrap";
 import HomePageTabImageNav from "@/components/GenericPagesComponents/HomePageTabImageNav";
 import TeachersTab from "@/components/GenericPagesComponents/TeachersTab";
 import CurriculumTab from "@/components/GenericPagesComponents/CurriculumTab";
